{"schema":"decodex/commit/1","summary":"{\"schema\":\"decodex/commit/1\",\"summary\":\"Split live adapter ELF runtime modules\",\"authority\":\"manual\"}","authority":"manual"}
